update : repackage controller, util and rename GameController.java
