+++ src/main/java/Main.java
@@ -11,10 +11,8 @@
  *
  */
 
-import controller.GameController;
-
 public class Main {
     public static void main(String[] args) {
-        new GameController().play();
+        new GameManager().play();
     }
 }
