+++ src/main/java/util/InputExceptionController.java
@@ -11,9 +11,8 @@
  *
  */
 
-package controller;
+package util;
 
-import view.Message;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
