+++ src/main/java/util/Message.java
@@ -11,7 +11,7 @@
  *
  */
 
-package view;
+package util;
 
 public enum Message {
     INPUT_PLAYER_NAME("게임에 참여할 사람의 이름을 입력하세요(쉼표 기준으로 분리)"),
