+++ src/main/java/view/InputView.java
@@ -16,7 +16,7 @@ package view;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
-import controller.InputExceptionController;
+import util.InputExceptionController;
 
 public class InputView {
     private final String SEPARATOR = ",";
